+++ TIGLCreator/src/TIGLCreatorWindow.cpp
@@ -1012,6 +1012,7 @@ void TIGLCreatorWindow::connectSignals()
     connect(modificatorModel, SIGNAL(configurationEdited()), this, SLOT(updateScene()));
     connect(modificatorModel, SIGNAL(configurationEdited()), this, SLOT(changeColorSaveButton()));
     connect(modificatorModel, SIGNAL(configurationEdited()), this, SLOT(dispatchLastSelectedItemOnConfigurationEdited()));
+    connect(modificatorModel, SIGNAL(configurationEdited()), this, SLOT(updateMenus()));
 
     connect(treeWidget, SIGNAL(newSelectedTreeItem(cpcr::CPACSTreeItem*)), modificatorModel, SLOT(dispatch(cpcr::CPACSTreeItem*)));
     connect(treeWidget, &CPACSTreeWidget::deleteSectionRequested, modificatorModel, &ModificatorModel::deleteSection);